+++ src/app/app-routing/app-routing.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+// App Components
+import { OpenLoansComponent } from '../views/open-loans/open-loans.component';
+import { ActiveLoansComponent } from '../active-loans/active-loans.component';
+import { LoanDetailComponent } from '../views/loan-detail/loan-detail.component';
+import { AddressComponent } from '../views/address/address.component';
+import { ProfileComponent } from '../views/profile/profile.component';
+import { NotFoundModule } from '../not-found/not-found.module';
+
+const appRoutes: Routes = [
+  { path: '', redirectTo: '/requests', pathMatch: 'full'},
+  { path: 'requests', component: OpenLoansComponent },
+  { path: 'activity', component: ActiveLoansComponent },
+  { path: 'loan/:id', component: LoanDetailComponent },
+  { path: 'address/:address', component: AddressComponent },
+  { path: 'profile', component: ProfileComponent },
+  { path: '404', component: NotFoundModule },
+  { path: '**',  redirectTo: '/404' },
+];
+
+@NgModule({
+  imports: [RouterModule.forRoot(appRoutes)],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }
